+++ app/app/src/main/java/com/github/livingwithhippos/unchained/base/model/network/ApiFactory.kt
@@ -7,8 +7,8 @@ import com.github.livingwithhippos.unchained.base.di.AuthRetrofit
 import com.github.livingwithhippos.unchained.downloaddetails.model.StreamingApi
 import com.github.livingwithhippos.unchained.downloaddetails.model.StreamingApiHelper
 import com.github.livingwithhippos.unchained.downloaddetails.model.StreamingApiHelperImpl
-import com.github.livingwithhippos.unchained.downloadlists.model.DownloadApiHelper
-import com.github.livingwithhippos.unchained.downloadlists.model.DownloadApiHelperImpl
+import com.github.livingwithhippos.unchained.lists.model.DownloadApiHelper
+import com.github.livingwithhippos.unchained.lists.model.DownloadApiHelperImpl
 import com.github.livingwithhippos.unchained.downloadlists.model.DownloadsApi
 import com.github.livingwithhippos.unchained.newdownload.model.TorrentApiHelper
 import com.github.livingwithhippos.unchained.newdownload.model.TorrentApiHelperImpl
