+++ app/app/src/main/java/com/github/livingwithhippos/unchained/downloadlists/viewmodel/DownloadListViewModel.kt
@@ -3,7 +3,6 @@ package com.github.livingwithhippos.unchained.downloadlists.viewmodel
 import androidx.hilt.Assisted
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -16,14 +15,9 @@ import com.github.livingwithhippos.unchained.base.model.repositories.Credentials
 import com.github.livingwithhippos.unchained.base.model.repositories.DownloadRepository
 import com.github.livingwithhippos.unchained.base.model.repositories.TorrentsRepository
 import com.github.livingwithhippos.unchained.downloadlists.model.DownloadItem
-import com.github.livingwithhippos.unchained.downloadlists.model.DownloadPagingSource
+import com.github.livingwithhippos.unchained.lists.model.DownloadPagingSource
 import com.github.livingwithhippos.unchained.lists.model.TorrentPagingSource
 import com.github.livingwithhippos.unchained.newdownload.model.TorrentItem
-import com.github.livingwithhippos.unchained.utilities.KEY_TOKEN
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.launch
 
 class DownloadListViewModel @ViewModelInject constructor(
     @Assisted private val savedStateHandle: SavedStateHandle,
