+++ app/app/src/main/java/com/github/livingwithhippos/unchained/lists/model/DownloadApiHelper.kt
@@ -1,5 +1,6 @@
-package com.github.livingwithhippos.unchained.downloadlists.model
+package com.github.livingwithhippos.unchained.lists.model
 
+import com.github.livingwithhippos.unchained.downloadlists.model.DownloadItem
 import retrofit2.Response
 
 interface DownloadApiHelper {
